rearange a bit before importing distributed test
